Strip extraneous new lines during rich text copy/paste.
This may cause things to look weird in the rich text box after paste,
but I don't know how to fix that part.

It might also affect the formatting of specially formatted text
with a lot of newlines.

I am only handling the chapter text rich text box right now,
because that's the one most likely to be dealing with copy/paste issues.