Fix getting work_type for AO3 import from user's default work_type
Previously, if a user had a default work type set, importing from AO3 would fail with: 'WorkType' object has no attribute 'first'. This was because work_type_mapping was set to the user's default_work_type, and we then tried to get the work type from this value as if it were a QuerySet.